Fix mypy CI failures: add pydantic to requirements and fix Python 3.8 typing compat
Add pydantic to requirements.txt so it is available during mypy checks,
bump typing-extensions for pydantic v2 compatibility, import Annotated
from typing_extensions for Python 3.8, update Optional type annotations
to match the new Pydantic model field types, and upgrade mypy to 1.5.1
(latest Python 3.8 compatible) since 0.782 crashes on pydantic v2.

+++ signal_ocean/voyages_market_data/models.py
@@ -152,7 +152,7 @@ class MatchedFixture(SignalBaseModel):
     cargo_group: Optional[str] = None
     quantity: Optional[int] = None
     quantity_buffer: Optional[int] = None
-    rate: Optional[int] = None
+    rate: Optional[float] = None
     rate_type: Optional[str] = None
     ballast_bonus_value: Optional[int] = None
     ballast_bonus_type: Optional[str] = None
